18.LLMs/02.po.extract/requirements.txt
flask==2.3.3
mysql-connector-python==8.1.0
openai==0.28.1
python-dotenv==1.0.0
pytesseract==0.3.10
pdf2image==1.16.3
gradio>=3.50.0
